Update HT_Change_work_environment.md
just an idea, it might be better to replace some of the full screen screenshots with only the parts of the window that are important to the point being made. Otherwise, it is sometimes difficult to see the part being highlighted. for example, the last screenshot on this page suggests looking at the main supervisor window, but the info in the screenshot is to small to be seen,